#12 - bieehoang-desktop

+++ store/models.py
@@ -6,7 +6,7 @@ class Product(models.Model):
     product_name = models.CharField(max_length=200, unique=True)
     slug = models.SlugField(max_length=200, unique=True)
     description = models.TextField(max_length=800, blank=True)
-    price = models.FloatField()
+    price = models.IntegerField()
     image = models.ImageField(upload_to='photos/products')
     stock = models.IntegerField()
     is_available = models.BooleanField(default=True)
